# backend-java/src/main/resources/static/openapi/visual-search-openapi.yaml
openapi: 3.0.3
info:
  title: Visual Search Engine API
  version: v1
  description: APIs for Authentication, Search, Search History and Bookmarks.
servers:
  - url: /visual-search/v1
    description: Local server
tags:
  - name: Authentication
  - name: Search
  - name: Search History
  - name: Bookmarks

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: quannh08
              email: quan@example.com
              password: StrongPassword@123
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterBaseResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '409':
          description: Username or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /auth/login:
    post:
      tags: [Authentication]
      summary: Login
      description: Returns an access token and sets refreshToken as an HttpOnly cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              usernameOrEmail: quannh08
              password: StrongPassword@123
      responses:
        '200':
          description: Login successfully
          headers:
            Set-Cookie:
              description: refreshToken HttpOnly cookie.
              schema:
                type: string
                example: refreshToken=<token>; HttpOnly; Secure; SameSite=Lax; Path=/visual-search/v1/auth
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginBaseResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '403':
          description: Account inactive or blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /auth/refresh-token:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Requires refreshToken HttpOnly cookie. Does not require Bearer access token.
      security: []
      parameters:
        - name: refreshToken
          in: cookie
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Access token refreshed successfully
          headers:
            Set-Cookie:
              description: Rotated refreshToken HttpOnly cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenBaseResponse'
        '401':
          description: Refresh token invalid, expired or revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      description: Requires Bearer access token. Clears refreshToken cookie.
      security:
        - bearerAuth: []
      parameters:
        - name: refreshToken
          in: cookie
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Logout successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutBaseResponse'
        '401':
          description: Unauthorized or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeBaseResponse'
        '401':
          description: Unauthorized or access token expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /auth/password:
    put:
      tags: [Authentication]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              currentPassword: StrongPassword@123
              newPassword: NewStrongPassword@456
              confirmNewPassword: NewStrongPassword@456
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordBaseResponse'
        '400':
          description: Validation error or password confirmation mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '401':
          description: Unauthorized or current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /search/image:
    post:
      tags: [Search]
      summary: Search by image
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageSearchBaseResponse'
        '400':
          description: Image required or invalid pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '415':
          description: Unsupported image type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '500':
          description: AI service or vector search error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /search/text:
    get:
      tags: [Search]
      summary: Search by text
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          example: red running shoes
        - name: mode
          in: query
          required: true
          schema:
            type: string
            enum: [semantic, ocr]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Text search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextSearchBaseResponse'
        '400':
          description: Query required, invalid mode or pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '500':
          description: AI, vector or OCR search error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /search-history:
    get:
      tags: [Search History]
      summary: Get search history list
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Search history list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchHistoryListBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
    delete:
      tags: [Search History]
      summary: Delete all search history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All search history deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAllSearchHistoryBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /search-history/{historyId}:
    get:
      tags: [Search History]
      summary: Get search history detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/HistoryIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Search history detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchHistoryDetailBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '404':
          description: Search history not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
    delete:
      tags: [Search History]
      summary: Delete one search history item
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/HistoryIdParam'
      responses:
        '200':
          description: Search history deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSearchHistoryBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '404':
          description: Search history not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /bookmarks:
    get:
      tags: [Bookmarks]
      summary: Get bookmarked images
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Bookmark list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookmarkListBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

  /bookmarks/{imageId}:
    post:
      tags: [Bookmarks]
      summary: Add bookmark
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      responses:
        '201':
          description: Bookmark created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBookmarkBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '409':
          description: Bookmark already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
    delete:
      tags: [Bookmarks]
      summary: Remove bookmark
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ImageIdParam'
      responses:
        '200':
          description: Bookmark removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteBookmarkBaseResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'
        '404':
          description: Bookmark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBaseResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
    HistoryIdParam:
      name: historyId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ImageIdParam:
      name: imageId
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    BaseResponse:
      type: object
      required: [success, data, error, timestamp]
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          nullable: true
          $ref: '#/components/schemas/ErrorResponse'
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      nullable: true
      properties:
        code:
          type: string
        message:
          type: string
    ErrorBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              example: false
            data:
              nullable: true
              example: null
            error:
              $ref: '#/components/schemas/ErrorResponse'

    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
    LoginRequest:
      type: object
      required: [usernameOrEmail, password]
      properties:
        usernameOrEmail:
          type: string
        password:
          type: string
          format: password
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword, confirmNewPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
        confirmNewPassword:
          type: string
          format: password

    UserRole:
      type: string
      enum: [ADMIN, USER]
    UserStatus:
      type: string
      enum: [ACTIVE, INACTIVE, BLOCKED]
    RegisterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
    LoginUserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          example: 86400
        user:
          $ref: '#/components/schemas/LoginUserResponse'
    RefreshTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          example: 86400
    MeResponse:
      $ref: '#/components/schemas/RegisterResponse'
    MessageResponse:
      type: object
      properties:
        message:
          type: string

    RegisterBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RegisterResponse'
    LoginBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoginResponse'
    RefreshTokenBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RefreshTokenResponse'
    MeBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/MeResponse'
    ChangePasswordBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/MessageResponse'
    LogoutBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/MessageResponse'

    SearchResultItem:
      type: object
      properties:
        imageId:
          type: integer
          format: int64
        originalFilename:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        similarityScore:
          type: number
          format: float
        rankPosition:
          type: integer
        width:
          type: integer
        height:
          type: integer
        mimeType:
          type: string
    ImageSearchResponse:
      type: object
      properties:
        searchId:
          type: integer
          format: int64
        searchType:
          type: string
          example: IMAGE_TO_IMAGE
        queryImageUrl:
          type: string
        processingTimeMs:
          type: integer
          format: int64
        content:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultItem'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    TextSearchResponse:
      allOf:
        - $ref: '#/components/schemas/ImageSearchResponse'
        - type: object
          properties:
            queryText:
              type: string
            mode:
              type: string
              enum: [semantic, ocr]
    ImageSearchBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ImageSearchResponse'
    TextSearchBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TextSearchResponse'

    SearchHistoryItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        searchType:
          type: string
        queryText:
          type: string
          nullable: true
        queryImageUrl:
          type: string
          nullable: true
        resultCount:
          type: integer
        processingTimeMs:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    SearchHistoryListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SearchHistoryItem'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    SearchHistoryDetailResponse:
      allOf:
        - $ref: '#/components/schemas/SearchHistoryItem'
        - type: object
          properties:
            results:
              $ref: '#/components/schemas/SearchHistoryResultPage'
    SearchHistoryResultPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SearchHistoryResultItem'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    SearchHistoryResultItem:
      type: object
      properties:
        imageId:
          type: integer
          format: int64
        originalFilename:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        similarityScore:
          type: number
          format: float
        rankPosition:
          type: integer
    SearchHistoryListBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SearchHistoryListResponse'
    SearchHistoryDetailBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SearchHistoryDetailResponse'
    DeleteSearchHistoryBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                deleted:
                  type: boolean
    DeleteAllSearchHistoryBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                deletedCount:
                  type: integer
                  format: int64

    BookmarkItem:
      type: object
      properties:
        bookmarkId:
          type: integer
          format: int64
        imageId:
          type: integer
          format: int64
        originalFilename:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        width:
          type: integer
        height:
          type: integer
        mimeType:
          type: string
        bookmarkedAt:
          type: string
          format: date-time
    BookmarkListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookmarkItem'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    BookmarkListBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BookmarkListResponse'
    CreateBookmarkBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                imageId:
                  type: integer
                  format: int64
                createdAt:
                  type: string
                  format: date-time
    DeleteBookmarkBaseResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                imageId:
                  type: integer
                  format: int64
                deleted:
                  type: boolean